requirements.txt
pyserial==3.5
paho-mqtt==1.6.1
fastcrc==0.2.1